requirements.txt
Pillow
pillow-heif
pillow-avif-plugin
numpy
matplotlib
jupyterlab
tqdm